chore: rename info.py to constants.py
