+++ src/app_type_dialog.py
@@ -18,7 +18,7 @@
 
 from gi.repository import Gtk, Adw
 
-from .info import rootdir
+from .constants import rootdir
 
 
 @Gtk.Template(resource_path=f"{rootdir}/ui/app_type_dialog.ui")
