+++ src/custom_css_group.py
@@ -18,7 +18,7 @@
 
 from gi.repository import Gtk, Adw
 
-from .info import rootdir
+from .constants import rootdir
 
 
 @Gtk.Template(resource_path=f"{rootdir}/ui/custom_css_group.ui")
