+++ src/palette_shades.py
@@ -18,7 +18,7 @@
 
 from gi.repository import Gtk, Gdk, Adw
 
-from .info import rootdir
+from .constants import rootdir
 
 
 @Gtk.Template(resource_path=f"{rootdir}/ui/palette_shades.ui")
