+++ src/window.py
@@ -34,7 +34,7 @@ from .option import AdwcustomizerOption
 from .app_type_dialog import AdwcustomizerAppTypeDialog
 from .custom_css_group import AdwcustomizerCustomCSSGroup
 from material_color_utilities_python import *
-from .info import rootdir, build_type
+from .constants import rootdir, build_type
 
 
 @Gtk.Template(resource_path=f"{rootdir}/ui/window.ui")
